requirements.txt
dawdreamer
numpy
scipy
soundfile
pyarrow
pyyaml
pandas
tqdm
pytest
